+++ frontend/public/environments/webnode.js
@@ -12,10 +12,6 @@ export default {
       'block-production': ['won-slots'],
     },
   },
-  sentry: {
-    dsn: 'https://69aba72a6290383494290cf285ab13b3@o4508216158584832.ingest.de.sentry.io/4508216160616528',
-    tracingOrigins: [],
-  },
   configs: [
     {
       name: 'Web Node',