+++ frontend/src/app/features/block-production/won-slots/table/block-production-won-slots-table.component.ts
@@ -23,8 +23,6 @@ import {
 } from '@shared/types/block-production/won-slots/block-production-won-slots-slot.type';
 import { BlockProductionWonSlotsSelectors } from '@block-production/won-slots/block-production-won-slots.state';
 import { BlockProductionWonSlotsActions } from '@block-production/won-slots/block-production-won-slots.actions';
-import { SentryService } from '@core/services/sentry.service';
-import { WebNodeService } from '@core/services/web-node.service';
 
 @Component({
   selector: 'mina-block-production-won-slots-table',
@@ -71,11 +69,7 @@ export class BlockProductionWonSlotsTableComponent
 
   private fromRoute: string;
 
-  constructor(
-    private router: Router,
-    private sentryService: SentryService,
-    private webnodeService: WebNodeService,
-  ) {
+  constructor(private router: Router) {
     super();
   }
 
@@ -86,19 +80,6 @@ export class BlockProductionWonSlotsTableComponent
     this.listenToRouteChange();
     this.listenToActiveSlotChange();
     this.listenToNodesChanges();
-
-    // this.select(BlockProductionWonSlotsSelectors.filteredSlots, (slots: BlockProductionWonSlotsSlot[]) => {
-    //   const blockProductionWonSlotsSlot = slots.find(d => d.message.includes('Confirm') || d.message.includes('Producing'));
-    //
-    //   if (blockProductionWonSlotsSlot?.globalSlot !== this.currentlyProducing?.globalSlot) {
-    //     if (!blockProductionWonSlotsSlot?.globalSlot) {
-    //       const block = slots.find(d => d.globalSlot === this.currentlyProducing?.globalSlot);
-    //       this.sentryService.updateProducedBlock(block, this.webnodeService.publicKey);
-    //     }
-    //     this.currentlyProducing = blockProductionWonSlotsSlot;
-    //   }
-    //   this.detect();
-    // }, filter((slots: BlockProductionWonSlotsSlot[]) => slots.length > 0));
   }
 
   protected override setupTable(): void {